pyproject.toml
[tool.poetry]
name = "sssd"
version = "0.0.16"
description = ""
authors = [
    "Yu-Ting Fan <ytfan.st07g@nctu.edu.tw>",
    "Joseph Wang <egpivo@gmail.com>",
    "Hsin-Cheng Huang <hchuang@stat.sinica.edu.tw>",
    "ShengLi Tzeng <slt.cmu@gmail.com>"
]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10.13"
h5py = "3.10.0"
numpy = "^1.24.4"
pandas = "2.2.1"
scikit-learn = "1.0.2"
scipy = "1.12.0"
torchmetrics = "0.8.2"
tqdm = "4.64.1"
pyyaml = "^6.0.1"
einops = "^0.4.1"
protobuf = "^3.20.1"
opt-einsum = "^3.3.0"
torch = [
    { markers = "sys_platform == 'darwin' and platform_machine == 'arm64'", url = "https://download.pytorch.org/whl/cpu/torch-2.2.0-cp310-none-macosx_11_0_arm64.whl" },
    { markers = "sys_platform == 'darwin' and platform_machine == 'x86_64'", url = "https://download.pytorch.org/whl/cpu/torch-2.2.0-cp310-none-macosx_10_9_x86_64.whl" },
    { markers = "sys_platform == 'linux'", url = "https://download.pytorch.org/whl/cu121/torch-2.2.0%2Bcu121-cp310-cp310-linux_x86_64.whl" }
]
statsmodels = "^0.14.1"
tensorboard = "^2.16.2"
gridstatus = "^0.27.0"
pykeops = "^2.2.3" # for Cauchy CUDA extension
boto3 = "^1.34.113"

[tool.poetry.group.dev.dependencies]
black = "^22.3.0"
isort = "^5.10.1"
mypy = "^0.971"
pydocstyle = { extras = ["toml"], version = "^6.1.1" }
pytest = "^7.1.3"
pytest-cov = "^4.1.0"
poethepoet = "^0.15.0"
coverage = { extras = ["toml"], version = "^6.4.4" }
pre-commit = "^2.20.0"
python-dotenv = "^1.0.0"
poetry-bumpversion = "^0.3.0"

[tool.poetry.group.notebook.dependencies]
ipykernel = "^6.29.4"
ipython = "^8.24.0"
matplotlib = "^3.8.4"
seaborn = "^0.13.2"
jinja2 = "^3.1.3"
jupyter = "^1.0.0"
jupyterlab = "^4.1.7"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
target-version = ['py310']

[tool.isort]
profile = "black"

[[tool.mypy.overrides]]
module = "cleo.*"
ignore_missing_imports = true
